refactor(checkable-input): convert checkable input to ts
