+++ src/__internal__/checkable-input/checkable-input-svg-wrapper.style.ts
