--- src/__internal__/checkable-input/hidden-checkable-input.d.ts
@@ -1,42 +1,0 @@
-import * as React from "react";
-
-export interface CommonHiddenCheckableInputProps
-  extends Omit<
-    React.InputHTMLAttributes<HTMLInputElement>,
-    "value" | "size" | "type"
-  > {
-  /** If true the Component will be focused when page load */
-  autoFocus?: boolean;
-  /** Checked state of the input */
-  checked?: boolean;
-  /** Input name */
-  name?: string;
-  /** OnChange event handler */
-  onChange?: (ev: React.ChangeEvent<HTMLInputElement>) => void;
-  /** OnFocus event handler */
-  onFocus?: (ev: React.FocusEvent<HTMLInputElement>) => void;
-  /** Blur event handler */
-  onBlur?: (ev: React.FocusEvent<HTMLInputElement>) => void;
-  /** OnMouseLeave event handler */
-  onMouseLeave?: (ev: React.MouseEvent<HTMLInputElement>) => void;
-  /** OnMouseEnter event handler */
-  onMouseEnter?: (ev: React.MouseEvent<HTMLInputElement>) => void;
-  /** Value of the input */
-  value?: string;
-  /** A callback to retrieve the input reference */
-  inputRef?: React.Ref<HTMLInputElement>;
-}
-
-export interface HiddenCheckableInputProps
-  extends CommonHiddenCheckableInputProps {
-  /** HTML type attribute of the input */
-  type: string;
-  /** Role attribute of the input */
-  role?: string;
-}
-
-declare function HiddenCheckableInput(
-  props: HiddenCheckableInputProps
-): JSX.Element;
-
-export default HiddenCheckableInput;
