+++ src/__internal__/checkable-input/hidden-checkable-input.style.ts
