--- src/__internal__/checkable-input/index.d.ts
@@ -1,5 +1,0 @@
-export {
-  default,
-  CheckableInputProps,
-  CommonCheckableInputProps,
-} from "./checkable-input";
