--- src/__internal__/checkable-input/index.js
@@ -1,1 +1,0 @@
-export { default } from "./checkable-input.component";
