+++ src/__internal__/checkable-input/index.ts
@@ -1,0 +1,5 @@
+export { default } from "./checkable-input.component";
+export type {
+  CheckableInputProps,
+  CommonCheckableInputProps,
+} from "./checkable-input.component";
